requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
newsapi-python==0.2.6
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
